- Moves sticker for beauty and cooking templates to new directories, fresh-and-bright and food-and-stuff respectively.
