+++ packages/stickers/src/food-and-stuff/arrow.js
