+++ packages/stickers/src/food-and-stuff/artichoke.js
