+++ packages/stickers/src/food-and-stuff/carrot.js
