+++ packages/stickers/src/food-and-stuff/index.js
