+++ packages/stickers/src/food-and-stuff/persimmon.js
