+++ packages/stickers/src/food-and-stuff/radish.js
