+++ packages/stickers/src/food-and-stuff/radishLarge.js
