+++ packages/stickers/src/fresh-and-bright/cta.js
