+++ packages/stickers/src/fresh-and-bright/heart.js
