+++ packages/stickers/src/fresh-and-bright/index.js
